refactor: 스케줄러 디렉토리 명 변경
